refactor: move LotteryResult
