+++ src/main/java/step2/OutOfMatchingBoundaryException.java
@@ -1,0 +1,4 @@
+package step2;
+
+class OutOfMatchingBoundaryException extends RuntimeException {
+}
